Rename run_X => X_command
