+++ libvast/vast/system/pcap_writer_command.hpp
@@ -11,8 +11,8 @@
  * contained in the LICENSE file.                                             *
  ******************************************************************************/
 
-#ifndef VAST_SYSTEM_RUN_WRITER_HPP
-#define VAST_SYSTEM_RUN_WRITER_HPP
+#ifndef VAST_SYSTEM_PCAP_WRITER_COMMAND_HPP
+#define VAST_SYSTEM_PCAP_WRITER_COMMAND_HPP
 
 #include <memory>
 #include <string>
@@ -22,50 +22,42 @@
 #include <caf/typed_actor.hpp>
 #include <caf/typed_event_based_actor.hpp>
 
+#include "vast/expression.hpp"
 #include "vast/logger.hpp"
 
-#include "vast/system/run_writer_base.hpp"
-#include "vast/system/sink.hpp"
+#include "vast/system/writer_command_base.hpp"
+#include "vast/system/writer_command_base.hpp"
+#include "vast/system/signal_monitor.hpp"
+#include "vast/system/source.hpp"
+#include "vast/system/tracker.hpp"
+
+#include "vast/format/pcap.hpp"
+
+#include "vast/concept/parseable/to.hpp"
+
+#include "vast/concept/parseable/vast/expression.hpp"
+#include "vast/concept/parseable/vast/schema.hpp"
 
 #include "vast/detail/make_io_stream.hpp"
 
 namespace vast::system {
 
-/// Default implementation for export sub-commands. Compatible with Bro and MRT
-/// formats.
+/// PCAP subcommant to `import`.
 /// @relates application
-template <class Writer>
-class run_writer : public run_writer_base {
+class pcap_writer_command : public writer_command_base {
 public:
-  run_writer(command* parent, std::string_view name)
-      : run_writer_base(parent, name),
-        output_("-"),
-        uds_(false) {
-    this->add_opt("write,w", "path to write events to", output_);
-    this->add_opt("uds,d", "treat -w as UNIX domain socket to connect to",
-                  uds_);
-  }
+  using super = writer_command_base;
+
+  pcap_writer_command(command* parent, std::string_view name);
 
 protected:
   expected<caf::actor> make_sink(caf::scoped_actor& self,
-                                 caf::message args) override {
-    CAF_LOG_TRACE(CAF_ARG(args));
-    using ostream_ptr = std::unique_ptr<std::ostream>;
-    if constexpr (std::is_constructible<Writer, ostream_ptr>::value) {
-      auto out = detail::make_output_stream(output_, uds_);
-      if (!out)
-        return out.error();
-      Writer writer{std::move(*out)};
-      return self->spawn(sink<Writer>, std::move(writer));
-    } else {
-      Writer writer;
-      return self->spawn(sink<Writer>, std::move(writer));
-    }
-  }
+                                 caf::message args) override;
 
 private:
   std::string output_;
   bool uds_;
+  unsigned flush_;
 };
 
 } // namespace vast::system
