+++ libvast/vast/system/writer_command.hpp
@@ -11,8 +11,8 @@
  * contained in the LICENSE file.                                             *
  ******************************************************************************/
 
-#ifndef VAST_SYSTEM_RUN_PCAP_READER_HPP
-#define VAST_SYSTEM_RUN_PCAP_READER_HPP
+#ifndef VAST_SYSTEM_RUN_WRITER_HPP
+#define VAST_SYSTEM_RUN_WRITER_HPP
 
 #include <memory>
 #include <string>
@@ -22,47 +22,50 @@
 #include <caf/typed_actor.hpp>
 #include <caf/typed_event_based_actor.hpp>
 
-#include "vast/expression.hpp"
 #include "vast/logger.hpp"
 
-#include "vast/system/run_reader_base.hpp"
-#include "vast/system/run_reader_base.hpp"
-#include "vast/system/signal_monitor.hpp"
-#include "vast/system/source.hpp"
-#include "vast/system/tracker.hpp"
-
-#include "vast/format/pcap.hpp"
-
-#include "vast/concept/parseable/to.hpp"
-
-#include "vast/concept/parseable/vast/expression.hpp"
-#include "vast/concept/parseable/vast/schema.hpp"
+#include "vast/system/writer_command_base.hpp"
+#include "vast/system/sink.hpp"
 
 #include "vast/detail/make_io_stream.hpp"
 
 namespace vast::system {
 
-/// PCAP subcommant to `import`.
+/// Default implementation for export sub-commands. Compatible with Bro and MRT
+/// formats.
 /// @relates application
-class run_pcap_reader : public run_reader_base {
+template <class Writer>
+class writer_command : public writer_command_base {
 public:
-  using super = run_reader_base;
-
-  run_pcap_reader(command* parent, std::string_view name);
+  writer_command(command* parent, std::string_view name)
+      : writer_command_base(parent, name),
+        output_("-"),
+        uds_(false) {
+    this->add_opt("write,w", "path to write events to", output_);
+    this->add_opt("uds,d", "treat -w as UNIX domain socket to connect to",
+                  uds_);
+  }
 
 protected:
-  expected<caf::actor> make_source(caf::scoped_actor& self,
-                                   caf::message args) override;
+  expected<caf::actor> make_sink(caf::scoped_actor& self,
+                                 caf::message args) override {
+    CAF_LOG_TRACE(CAF_ARG(args));
+    using ostream_ptr = std::unique_ptr<std::ostream>;
+    if constexpr (std::is_constructible<Writer, ostream_ptr>::value) {
+      auto out = detail::make_output_stream(output_, uds_);
+      if (!out)
+        return out.error();
+      Writer writer{std::move(*out)};
+      return self->spawn(sink<Writer>, std::move(writer));
+    } else {
+      Writer writer;
+      return self->spawn(sink<Writer>, std::move(writer));
+    }
+  }
 
 private:
-  std::string input;
-  std::string schema_file;
-  bool uds;
-  uint64_t flow_max;
-  unsigned flow_age;
-  unsigned flow_expiry;
-  size_t cutoff;
-  int64_t pseudo_realtime;
+  std::string output_;
+  bool uds_;
 };
 
 } // namespace vast::system
